(wip) switch to secondary texture approach: not yet working

+++ src/mbgl/shader/extrusion_texture_shader.hpp
@@ -13,11 +13,10 @@ namespace mbgl {
         void bind(GLbyte *offset) final;
 
         UniformMatrix<4>                u_matrix   = {"u_matrix", *this};
-        Uniform<GLfloat>                u_xdim     = {"u_xdim", *this};
-        Uniform<GLfloat>                u_ydim     = {"u_ydim", *this};
+        Uniform<GLint>                  u_xdim     = {"u_xdim", *this};
+        Uniform<GLint>                  u_ydim     = {"u_ydim", *this};
         Uniform<GLfloat>                u_opacity  = {"u_opacity", *this};
         Uniform<GLint>                  u_texture  = {"u_texture", *this};
-        // TODO is this right or GLbyte? there are no other uniform ints in the shaders
     };
     
 } // namespace mbgl
